refactor(HeaderComponent): fixed to show kingdomName coming from API
