+++ Tribes/src/app/header.service.ts
@@ -1,4 +1,5 @@
-import { KingdomNameResponse } from './kingdomNameResponse';
+import { KingdomResponse } from './kingdom-response';
+import { ROOT_URL } from './constants';
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpResponse } from '@angular/common/http';
 import { Observable } from 'rxjs';
@@ -8,11 +9,9 @@ import { Observable } from 'rxjs';
 })
 export class HeaderService {
 
-  rootUrl = 'http://virtserver.swaggerhub.com/szutsj/godess-tribes-backend/1.0.0';
-
   constructor(private http: HttpClient) { }
 
-  getDataFromBackend(): Observable<KingdomNameResponse> {
-    return this.http.get<KingdomNameResponse>(this.rootUrl + '/kingdom');
+  getDataFromBackend(): Observable<KingdomResponse> {
+    return this.http.get<KingdomResponse>(ROOT_URL + '/kingdom');
   }
 }
