+++ Tribes/src/app/header/header.component.html
@@ -7,7 +7,7 @@
           <a class="nav-link"><img src="../../assets/images/strategyIcon.png"></a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" routerLink="/kingdom">{{name}}</a>
+          <a class="nav-link" routerLink="/kingdom">{{kingdomName}}</a>
         </li>
       </ul>
     </div>
