--- Tribes/src/app/kingdomNameResponse.ts
@@ -1,3 +1,0 @@
-export class KingdomNameResponse {
-   name: string;
-}
