+++ Tribes/src/app/location.ts
@@ -1,0 +1,5 @@
+export class Location {
+
+    xCoord: number;
+    yCoord: number;
+}
