+++ Tribes/src/app/soldier.ts
@@ -1,0 +1,7 @@
+export class Soldier {
+
+    id: number;
+    level: number;
+    startedAt: Date;
+    finsihedAd: Date;
+}
